requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
